alarm-consumer/requirements.txt
setuptools==60.0.0
flower==2.0.1
click==7.1.2
colorlog==6.6.0
ruamel.yaml==0.17.1
kubernetes==25.3.0
gevent==21.1.2
gunicorn==20.1.0
mysqlclient==1.4.6
python-dotenv==0.17.1
requests==2.25.1
uuid==1.30

celery==5.1.2
cryptography==2.4.2
PyMySQL==0.9.3
elasticsearch==8.3.3
ansible==2.9.27
simplejson==3.19.2
redis==4.3.3